f32 -> f64

+++ bevy_rapier2d/examples/contact_filter2.rs
@@ -81,7 +81,7 @@ pub fn setup_physics(mut commands: Commands) {
     let rad = 5.0;
 
     let shift = rad * 2.0;
-    let centerx = shift * (num as f32) / 2.0;
+    let centerx = shift * (num as f64) / 2.0;
     let centery = shift / 2.0;
     let mut group_id = 0;
     let tags = [CustomFilterTag::GroupA, CustomFilterTag::GroupB];
@@ -89,8 +89,8 @@ pub fn setup_physics(mut commands: Commands) {
 
     for i in 0..num {
         for j in 0usize..num * 5 {
-            let x = (i as f32 + j as f32 * 0.2) * shift - centerx;
-            let y = j as f32 * shift + centery + 20.0;
+            let x = (i as f64 + j as f64 * 0.2) * shift - centerx;
+            let y = j as f64 * shift + centery + 20.0;
             group_id += 1;
 
             commands.spawn((
