+++ bevy_rapier2d/examples/custom_system_setup2.rs
@@ -144,15 +144,15 @@ pub fn setup_physics(mut commands: Commands) {
     let rad = 10.0;
 
     let shift = rad * 2.0 + rad;
-    let centerx = shift * (num / 2) as f32;
+    let centerx = shift * (num / 2) as f64;
     let centery = shift / 2.0;
 
-    let mut offset = -(num as f32) * (rad * 2.0 + rad) * 0.5;
+    let mut offset = -(num as f64) * (rad * 2.0 + rad) * 0.5;
 
     for j in 0usize..20 {
         for i in 0..num {
-            let x = i as f32 * shift - centerx + offset;
-            let y = j as f32 * shift + centery + 30.0;
+            let x = i as f64 * shift - centerx + offset;
+            let y = j as f64 * shift + centery + 30.0;
 
             commands.spawn((
                 TransformBundle::from(Transform::from_xyz(x, y, 0.0)),
@@ -161,6 +161,6 @@ pub fn setup_physics(mut commands: Commands) {
             ));
         }
 
-        offset -= 0.05 * rad * (num as f32 - 1.0);
+        offset -= 0.05 * rad * (num as f64 - 1.0);
     }
 }
