+++ bevy_rapier2d/examples/multiple_colliders2.rs
@@ -36,18 +36,18 @@ pub fn setup_physics(mut commands: Commands) {
      * Create the cubes
      */
     let num = 4;
-    let rad = 2.0;
+    let rad = 2.0f64;
 
     let shift = rad * 4.0 + rad;
-    let centerx = shift * (num / 2) as f32;
+    let centerx = shift * (num / 2) as f64;
     let centery = shift / 2.0;
 
-    let mut offset = -(num as f32) * (rad * 2.0 + rad) * 0.5;
+    let mut offset = -(num as f64) * (rad * 2.0 + rad) * 0.5;
 
     for j in 0usize..20 {
         for i in 0..num {
-            let x = i as f32 * shift * 5.0 - centerx + offset;
-            let y = j as f32 * (shift * 5.0) + centery + 3.0;
+            let x = i as f64 * shift * 5.0 - centerx + offset;
+            let y = j as f64 * (shift * 5.0) + centery + 3.0;
 
             commands
                 .spawn((
@@ -67,6 +67,6 @@ pub fn setup_physics(mut commands: Commands) {
                 });
         }
 
-        offset -= 0.05 * rad * (num as f32 - 1.0);
+        offset -= 0.05 * rad * (num as f64 - 1.0);
     }
 }
