+++ src/dynamics/rigid_body.rs
@@ -62,7 +62,7 @@ pub struct Velocity {
     pub linvel: Vect,
     /// The angular velocity of the rigid-body.
     #[cfg(feature = "dim2")]
-    pub angvel: f32,
+    pub angvel: f64,
     /// The angular velocity of the rigid-body.
     #[cfg(feature = "dim3")]
     pub angvel: Vect,
@@ -84,7 +84,7 @@ impl Velocity {
 
     /// Initialize a velocity with the given angular velocity, and a linear velocity of zero.
     #[cfg(feature = "dim2")]
-    pub fn angular(angvel: f32) -> Self {
+    pub fn angular(angvel: f64) -> Self {
         Self {
             angvel,
             ..Self::default()
@@ -108,7 +108,7 @@ pub enum AdditionalMassProperties {
     /// This mass will be added to the rigid-body. The rigid-body’s total
     /// angular inertia tensor (obtained from its attached colliders) will
     /// be scaled accordingly.
-    Mass(f32),
+    Mass(f64),
     /// These mass properties will be added to the rigid-body.
     MassProperties(MassProperties),
 }
@@ -139,10 +139,10 @@ pub struct MassProperties {
     /// The center of mass of a rigid-body expressed in its local-space.
     pub local_center_of_mass: Vect,
     /// The mass of a rigid-body.
-    pub mass: f32,
+    pub mass: f64,
     /// The principal angular inertia of the rigid-body.
     #[cfg(feature = "dim2")]
-    pub principal_inertia: f32,
+    pub principal_inertia: f64,
     /// The principal vectors of the local angular inertia tensor of the rigid-body.
     #[cfg(feature = "dim3")]
     pub principal_inertia_local_frame: crate::math::Rot,
@@ -154,7 +154,7 @@ pub struct MassProperties {
 impl MassProperties {
     /// Converts these mass-properties to Rapier’s `MassProperties` structure.
     #[cfg(feature = "dim2")]
-    pub fn into_rapier(self, physics_scale: f32) -> rapier::dynamics::MassProperties {
+    pub fn into_rapier(self, physics_scale: f64) -> rapier::dynamics::MassProperties {
         rapier::dynamics::MassProperties::new(
             (self.local_center_of_mass / physics_scale).into(),
             self.mass,
@@ -165,7 +165,7 @@ impl MassProperties {
 
     /// Converts these mass-properties to Rapier’s `MassProperties` structure.
     #[cfg(feature = "dim3")]
-    pub fn into_rapier(self, physics_scale: f32) -> rapier::dynamics::MassProperties {
+    pub fn into_rapier(self, physics_scale: f64) -> rapier::dynamics::MassProperties {
         rapier::dynamics::MassProperties::with_principal_inertia_frame(
             (self.local_center_of_mass / physics_scale).into(),
             self.mass,
@@ -175,7 +175,7 @@ impl MassProperties {
     }
 
     /// Converts Rapier’s `MassProperties` structure to `Self`.
-    pub fn from_rapier(mprops: rapier::dynamics::MassProperties, physics_scale: f32) -> Self {
+    pub fn from_rapier(mprops: rapier::dynamics::MassProperties, physics_scale: f64) -> Self {
         #[allow(clippy::useless_conversion)] // Need to convert if dim3 enabled
         Self {
             mass: mprops.mass(),
@@ -228,7 +228,7 @@ pub struct ExternalForce {
     pub force: Vect,
     /// The angular torque applied to the rigid-body.
     #[cfg(feature = "dim2")]
-    pub torque: f32,
+    pub torque: f64,
     /// The angular torque applied to the rigid-body.
     #[cfg(feature = "dim3")]
     pub torque: Vect,
@@ -300,7 +300,7 @@ pub struct ExternalImpulse {
     pub impulse: Vect,
     /// The angular impulse applied to the rigid-body.
     #[cfg(feature = "dim2")]
-    pub torque_impulse: f32,
+    pub torque_impulse: f64,
     /// The angular impulse applied to the rigid-body.
     #[cfg(feature = "dim3")]
     pub torque_impulse: Vect,
@@ -370,7 +370,7 @@ impl SubAssign for ExternalImpulse {
 /// applied to this rigid-body.
 #[derive(Copy, Clone, Debug, PartialEq, Component, Reflect, FromReflect)]
 #[reflect(Component, PartialEq)]
-pub struct GravityScale(pub f32);
+pub struct GravityScale(pub f64);
 
 impl Default for GravityScale {
     fn default() -> Self {
@@ -425,9 +425,9 @@ impl Dominance {
 #[reflect(Component, PartialEq)]
 pub struct Sleeping {
     /// The threshold linear velocity bellow which the body can fall asleep.
-    pub linear_threshold: f32,
+    pub linear_threshold: f64,
     /// The angular linear velocity bellow which the body can fall asleep.
-    pub angular_threshold: f32,
+    pub angular_threshold: f64,
     /// Is this body sleeping?
     pub sleeping: bool,
 }
@@ -459,9 +459,9 @@ impl Default for Sleeping {
 pub struct Damping {
     // TODO: rename these to "linear" and "angular"?
     /// Damping factor for gradually slowing down the translational motion of the rigid-body.
-    pub linear_damping: f32,
+    pub linear_damping: f64,
     /// Damping factor for gradually slowing down the angular motion of the rigid-body.
-    pub angular_damping: f32,
+    pub angular_damping: f64,
 }
 
 impl Default for Damping {
@@ -479,14 +479,14 @@ impl Default for Damping {
 #[derive(Copy, Clone, Debug, Default, PartialEq, Component)]
 pub struct TransformInterpolation {
     /// The starting point of the interpolation.
-    pub start: Option<Isometry<f32>>,
+    pub start: Option<Isometry<f64>>,
     /// The end point of the interpolation.
-    pub end: Option<Isometry<f32>>,
+    pub end: Option<Isometry<f64>>,
 }
 
 impl TransformInterpolation {
     /// Interpolates between the start and end positions with `t` in the range `[0..1]`.
-    pub fn lerp_slerp(&self, t: f32) -> Option<Isometry<f32>> {
+    pub fn lerp_slerp(&self, t: f64) -> Option<Isometry<f64>> {
         if let (Some(start), Some(end)) = (self.start, self.end) {
             Some(start.lerp_slerp(&end, t))
         } else {
